7. Refactor project code and tests to your liking and best effort. Focus on QuizController and QuizService.
