+++ src/main/java/es/nitaur/api/quiz/dto/QuizAnswerDTO.java
@@ -1,0 +1,9 @@
+package es.nitaur.api.quiz.dto;
+
+import lombok.Data;
+
+@Data
+public class QuizAnswerDTO {
+    private Long id;
+    private String answer;
+}
