+++ src/main/java/es/nitaur/api/quiz/dto/QuizFormDTO.java
@@ -1,0 +1,10 @@
+package es.nitaur.api.quiz.dto;
+
+import lombok.Data;
+
+@Data
+public class QuizFormDTO {
+    private Long id;
+    private String language;
+    private QuizSectionDTO section;
+}
