+++ src/main/java/es/nitaur/persistence/repository/QuizFormRepository.java
@@ -1,6 +1,6 @@
-package es.nitaur.repository;
+package es.nitaur.persistence.repository;
 
-import es.nitaur.QuizForm;
+import es.nitaur.persistence.model.quiz.QuizForm;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
 import org.springframework.stereotype.Repository;
