+++ src/main/java/es/nitaur/persistence/repository/QuizRepository.java
@@ -1,0 +1,10 @@
+package es.nitaur.persistence.repository;
+
+import es.nitaur.persistence.model.quiz.Quiz;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface QuizRepository extends JpaRepository<Quiz, Long> {
+
+}
