--- src/main/java/es/nitaur/repository/QuizQuestionRepository.java
@@ -1,15 +1,0 @@
-package es.nitaur.repository;
-
-import es.nitaur.QuizQuestion;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Lock;
-import org.springframework.stereotype.Repository;
-
-import javax.persistence.LockModeType;
-
-@Repository
-public interface QuizQuestionRepository extends JpaRepository<QuizQuestion, Long> {
-
-    @Lock(LockModeType.PESSIMISTIC_WRITE)
-    QuizQuestion findOne(Long id);
-}
