--- src/main/java/es/nitaur/repository/QuizRepository.java
@@ -1,10 +1,0 @@
-package es.nitaur.repository;
-
-import es.nitaur.Quiz;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuizRepository extends JpaRepository<Quiz, Long> {
-
-}
